defer fetching non criticals
