+++ frontend/src/App.vue
@@ -1,11 +1,15 @@
 <script setup lang="ts">
 import Notification from "@/components/Notification.vue";
 import api from "@/services/api/index";
+import userApi from "@/services/api/user";
+import platformApi from "@/services/api/platform";
 import socket from "@/services/socket";
 import storeConfig from "@/stores/config";
 import storeGalleryFilter from "@/stores/galleryFilter";
 import storeHeartbeat from "@/stores/heartbeat";
 import storeRoms, { type Rom } from "@/stores/roms";
+import storePlatforms from "@/stores/platforms";
+import storeAuth from "@/stores/auth";
 import storeScanning from "@/stores/scanning";
 import type { Events } from "@/types/emitter";
 import { normalizeString } from "@/utils";
@@ -24,6 +28,8 @@ const emitter = inject<Emitter<Events>>("emitter");
 // Props
 const heartbeat = storeHeartbeat();
 const configStore = storeConfig();
+const auth = storeAuth();
+const platformsStore = storePlatforms();
 
 socket.on(
   "scan:scanning_platform",
@@ -93,16 +99,41 @@ onBeforeUnmount(() => {
   socket.off("scan:scanning_rom");
   socket.off("scan:done");
   socket.off("scan:done_ko");
+
+  document.removeEventListener("network-quiesced", fetchHomeData);
 });
 
-onBeforeMount(() => {
+function fetchHomeData() {
+  document.removeEventListener("network-quiesced", fetchHomeData);
+
   api.get("/heartbeat").then(({ data: data }) => {
     heartbeat.set(data);
   });
+
   api.get("/config").then(({ data: data }) => {
     configStore.set(data);
   });
-});
+
+  platformApi
+    .getPlatforms()
+    .then(({ data: platforms }) => {
+      platformsStore.set(platforms);
+    })
+    .catch((error) => {
+      console.error(error);
+    });
+
+  userApi
+    .fetchCurrentUser()
+    .then(({ data: user }) => {
+      auth.setUser(user);
+    })
+    .catch((error) => {
+      console.error(error);
+    });
+}
+
+document.addEventListener("network-quiesced", fetchHomeData)
 </script>
 
 <template>
