+++ frontend/src/views/Home.vue
@@ -20,14 +20,12 @@ import DeleteUserDialog from "@/components/Dialog/User/DeleteUser.vue";
 import EditUserDialog from "@/components/Dialog/User/EditUser.vue";
 import Drawer from "@/components/Drawer/Base.vue";
 import platformApi from "@/services/api/platform";
-import userApi from "@/services/api/user";
-import storeAuth from "@/stores/auth";
 import storePlatforms from "@/stores/platforms";
 import storeScanning from "@/stores/scanning";
 import type { Events } from "@/types/emitter";
 import type { Emitter } from "mitt";
 import { storeToRefs } from "pinia";
-import { inject, onBeforeUnmount, ref, nextTick } from "vue";
+import { inject, ref } from "vue";
 import { useDisplay } from "vuetify";
 
 // Props
@@ -35,7 +33,6 @@ const { mdAndDown } = useDisplay();
 const scanningStore = storeScanning();
 const { scanning } = storeToRefs(scanningStore);
 const platformsStore = storePlatforms();
-const auth = storeAuth();
 const refreshView = ref(0);
 
 // Event listeners bus
@@ -46,34 +43,6 @@ emitter?.on("refreshDrawer", async () => {
 });
 emitter?.on("refreshView", async () => {
   refreshView.value = refreshView.value + 1;
-});
-
-function fetchHomeData() {
-  document.removeEventListener("network-quiesced", fetchHomeData);
-
-  platformApi
-    .getPlatforms()
-    .then(({ data: platforms }) => {
-      platformsStore.set(platforms);
-    })
-    .catch((error) => {
-      console.error(error);
-    });
-
-  userApi
-    .fetchCurrentUser()
-    .then(({ data: user }) => {
-      auth.setUser(user);
-    })
-    .catch((error) => {
-      console.error(error);
-    });
-}
-
-document.addEventListener("network-quiesced", fetchHomeData);
-
-onBeforeUnmount(() => {
-  document.removeEventListener("network-quiesced", fetchHomeData);
 });
 </script>
